+++ NewPlatform.Flexberry.ORM.ODataService/Model/DataObjectEdmModel.cs
@@ -178,7 +178,7 @@
         private void BuildOwnProperties(EdmEntityType edmEntityType, Type dataObjectType)
         {
             var settings = _metadata[dataObjectType];
-            var keyPropertyType = settings.KeyType == null ? null : EdmTypeMap.GetEdmPrimitiveType(settings.KeyType);
+            var keyPropertyType = settings.KeyType == null ? null : EdmTypeMap.GetEdmPrimitiveType(settings.KeyType, EdmModelBuilder?.AdditionalMapping);
             IEnumerable<PropertyInfo> ownProperties = settings.OwnProperties;
             foreach (var propertyInfo in ownProperties)
             {
@@ -213,7 +213,7 @@
                     this.SetAnnotationValue(edmProp, new ClrPropertyInfoAnnotation(propertyInfo));
                 }
 
-                IEdmPrimitiveType edmPrimitiveType = EdmTypeMap.GetEdmPrimitiveType(propertyType);
+                IEdmPrimitiveType edmPrimitiveType = EdmTypeMap.GetEdmPrimitiveType(propertyType, EdmModelBuilder?.AdditionalMapping);
                 if (edmPrimitiveType != null)
                 {
                     EdmStructuralProperty edmProp = edmEntityType.AddStructuralProperty(GetEntityPropertyName(propertyInfo), edmPrimitiveType.PrimitiveKind);
@@ -681,7 +681,7 @@
             {
                 Type paramType = function.ParametersTypes[parameter];
                 IEdmEnumType enumType = GetEdmEnumType(paramType);
-                var edmParameterType = EdmTypeMap.GetEdmPrimitiveType(paramType);
+                var edmParameterType = EdmTypeMap.GetEdmPrimitiveType(paramType, EdmModelBuilder?.AdditionalMapping);
                 if (edmParameterType != null)
                 {
                     edmFunction.AddParameter(parameter, new EdmPrimitiveTypeReference(edmParameterType, false));
@@ -803,7 +803,7 @@
 
         private IEdmTypeReference GetNotEntityTypeReference(Type clrType)
         {
-            IEdmType edmType = EdmTypeMap.GetEdmPrimitiveType(clrType);
+            IEdmType edmType = EdmTypeMap.GetEdmPrimitiveType(clrType, EdmModelBuilder?.AdditionalMapping);
             if (edmType != null)
             {
                 return new EdmPrimitiveTypeReference(edmType as IEdmPrimitiveType, false);
@@ -841,10 +841,10 @@
         private void RegisterGeoIntersectsFunction()
         {
             EdmFunction edmFunction;
-            var edmReturnType = EdmTypeMap.GetEdmPrimitiveType(typeof(bool));
+            var edmReturnType = EdmTypeMap.GetEdmPrimitiveType(typeof(bool), EdmModelBuilder?.AdditionalMapping);
             edmFunction = new EdmFunction("geo", "intersects", new EdmPrimitiveTypeReference(edmReturnType, false), false, null, true);
             AddElement(edmFunction);
-            var edmParameterType = EdmTypeMap.GetEdmPrimitiveType(typeof(Geography));
+            var edmParameterType = EdmTypeMap.GetEdmPrimitiveType(typeof(Geography), EdmModelBuilder?.AdditionalMapping);
             edmFunction.AddParameter("geography1", new EdmPrimitiveTypeReference(edmParameterType, false));
             edmFunction.AddParameter("geography2", new EdmPrimitiveTypeReference(edmParameterType, false));
         }
@@ -852,10 +852,10 @@
         private void RegisterGeomIntersectsFunction()
         {
             EdmFunction edmFunction;
-            var edmReturnType = EdmTypeMap.GetEdmPrimitiveType(typeof(bool));
+            var edmReturnType = EdmTypeMap.GetEdmPrimitiveType(typeof(bool), EdmModelBuilder?.AdditionalMapping);
             edmFunction = new EdmFunction("geom", "intersects", new EdmPrimitiveTypeReference(edmReturnType, false), false, null, true);
             AddElement(edmFunction);
-            var edmParameterType = EdmTypeMap.GetEdmPrimitiveType(typeof(Geometry));
+            var edmParameterType = EdmTypeMap.GetEdmPrimitiveType(typeof(Geometry), EdmModelBuilder?.AdditionalMapping);
             edmFunction.AddParameter("geometry1", new EdmPrimitiveTypeReference(edmParameterType, false));
             edmFunction.AddParameter("geometry2", new EdmPrimitiveTypeReference(edmParameterType, false));
         }